reorganize pages and routes
