+++ gui/pages/Splash.js
