+++ gui/pages/session/SessionLobby.js
@@ -1,15 +1,15 @@
 const m = require('mithril');
 const classNames = require('classnames');
 
-import Page from '../models/Page.js';
-import Button from "../components/templates/Button.js";
-import Body from "../components/Body.js";
-import scenarioManifestStore from "../stores/ScenarioManifest.store.js";
-import scenarioDefinitionStore from "../stores/ScenarioDefinition.store.js";
-import TitleBar from "../components/templates/TitleBar.js";
+import Page from '../../models/Page.js';
+import Button from "../../components/templates/Button.js";
+import Body from "../../components/Body.js";
+import scenarioManifestStore from "../../stores/ScenarioManifest.store.js";
+import scenarioDefinitionStore from "../../stores/ScenarioDefinition.store.js";
+import TitleBar from "../../components/templates/TitleBar.js";
 
 
-export const page = new Page('/lobby', (initialVnode) => {
+export const page = new Page('/session/lobby', (initialVnode) => {
     const {loadScenarioManifestRegistry} = scenarioManifestStore;
     const {setActiveScenarioDefinition} = scenarioDefinitionStore;
     loadScenarioManifestRegistry().then(() => m.redraw());
@@ -24,7 +24,7 @@ export const page = new Page('/lobby', (initialVnode) => {
     const handleStartBattle = async (manifest) => {
         if (selectedScenario) {
             await setActiveScenarioDefinition(manifest);
-            m.route.set('/session');
+            m.route.set('/scenario');
         }
     }
